refactor(ssr): move `MutationTracker` to own file
